Change the unsigned long to portable types
The unsigned long size in gcc on 64-bit Linux is 64-bit while on MinGW
on 64-bit Windows is 32-bit (compiled with MinGW). This not only cause
compilation error but also failing tests on MinGW.

+++ m4ri/misc.h
@@ -418,7 +418,7 @@ static inline word m4ri_spread_bits(word const from, rci_t *const Q, int const l
  * \param n
  */
 
-#define __M4RI_ALIGNMENT(addr, n) (((unsigned long)(addr)) % (n))
+#define __M4RI_ALIGNMENT(addr, n) (((uintptr_t)(addr)) % (n))
 
 /**
  * \brief Test for gcc >= maj.min, as per __GNUC_PREREQ in glibc
